Agents/newssummarizer.py
from dotenv import load_dotenv

load_dotenv()
from langchain_community.tools.tavily_search import TavilySearchResults
from langchain_mistralai import ChatMistralAI
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import StrOutputParser

search_tool = TavilySearchResults(max_result=5)

llm = ChatMistralAI(model="mistral-small-2506")

prompt = ChatPromptTemplate.from_template(
  """
  You are a helpful assistant who summarizes news articles. You will be given a news article and you need to summarize it in a few sentences. The summary should capture the main points of the article and be easy to understand in clear bullet points.

  {news}
  """
)

chain = prompt | llm | StrOutputParser()

news_result=search_tool.run("Latest AI news of 2026")

result = chain.invoke({"news": news_result})
print(result)